RustConnection: Move XID generation to a dedicated helper
This should not have any behavioural changes, besides the extra RefCell.
